issue #279 Passphrase setup for EKM keys (#434)
* Handled empty or not decrypted keys while searching EKM keys;

* Added passphrase setup if any key present in EKM;

* Created abstract view controller for passphrase creation;
Fixed PR comments;

* fix bug when more than one key is used

* Saving passphrase in memory for EKM flow;
